requirements.txt
streamlit
google-cloud-firestore
streamlit-authenticator